Re-run code cells after deleting pickle
As suggested by @WuStangDan on https://waffle.io/udacity/sdc-issue-reports/cards/581e0df50be0b13200c4c11e,
This commit is the result of addressing a low cap in the accuracy.

Previously, the accuracy I was achieving was limited to 0.1 or so.
After this fix, as shown in the notebook itself, I was able to achieve
accuracy of 0.86 or so.

The pickle file itself is not contained in this notebook, as it is
about 500MB in size.